[fixup|fabric] Fix remove overlapping shards behavior
Previously overlapping shard removal was not working properly. If a smaller
copy of a shard returned, it would have incorrectly killed a larger copy that
was still waiting for a response.

The improvement is two-fold:

1) Fix the behavior so that workers are killed only after there is a full ring
present. This is happening in `fabric_ring:handle_response/4` function. There,
instead of re-using the same waiting workers fabric dict, build a new response
data structure `[{Range, Shard, ResponseVal}, ...]`. As responses come in,
workers are removed from the waiting dict and added to response. Then, if there
is a full ring, the rest of the waiting workers are killed and responses (and
workers) which participated in the ring are returned. In case of
fabric_streams, added a new `ready` worker list to its stream_acc accumulator
record.

2) Abstract the common pattern of gathering all the response until a ring is
formed, handling worker and node failures as well. This reduced quite a bit of
boilerplate code in fabric_doc_count, fabric_streams and other module.

(Thanks for Paul J. Davis for helping identify the issue and providing the
handle_response snippet)