[rexi] Optimize rexi:kill with rexi:kill_all
After the fabric ring updates, it turns out workers are killed in batches every
time, as opposed to individually as responses come in. So switch to using
rex:kill_all instead of rexi:kill. This means that in cases of requests that
needed responses for the full ring, ther were (N - 1) * Q dist messages sent
and now only (N - 1) messages are sent.